Rename check-and-save/tx to check-tx
